+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## UNRELEASED
+### Added
+- Add `timeout` param to `Cassette.st`, `Cassette.tgt` and `Cassette.validate`
 
 ## [1.4.1] - 2018-05-27
 ### Added